ddl: migrate test-infra to testify for testDBSuite1 and testDBSuite2
